+++ lib/savon/error.rb
@@ -1,0 +1,6 @@
+module Savon
+
+  # Base class for Savon errors.
+  class Error < RuntimeError; end
+
+end