+++ spec/savon/soap/response_spec.rb
@@ -3,24 +3,44 @@ require "spec_helper"
 describe Savon::SOAP::Response do
 
   describe ".new" do
-    it "raises a Savon::SOAPFault in case of a SOAP fault" do
-      lambda { new_response :body => ResponseFixture.soap_fault }.should raise_error(Savon::SOAPFault)
+    it "should raise a Savon::SOAP::Fault in case of a SOAP fault" do
+      lambda { soap_fault_response }.should raise_error(Savon::SOAP::Fault)
     end
 
-    it "does not raise a Savon::SOAPFault in case the default is turned off" do
+    it "should not raise a Savon::SOAP::Fault in case the default is turned off" do
       Savon.raise_errors = false
-      new_response :body => ResponseFixture.soap_fault
+      lambda { soap_fault_response }.should_not raise_error(Savon::SOAP::Fault)
       Savon.raise_errors = true
     end
 
-    it "raises a Savon::HTTPError in case of an HTTP error" do
-      lambda { new_response :code => 500 }.should raise_error(Savon::HTTPError)
+    it "should raise a Savon::HTTP::Error in case of an HTTP error" do
+      lambda { soap_response :code => 500 }.should raise_error(Savon::HTTP::Error)
     end
 
-    it "does not raise a Savon::HTTPError in case the default is turned off" do
+    it "should not raise a Savon::HTTP::Error in case the default is turned off" do
       Savon.raise_errors = false
-      new_response :code => 500
+      soap_response :code => 500
       Savon.raise_errors = true
+    end
+  end
+
+  describe "#success?" do
+    around do |example|
+      Savon.raise_errors = false
+      example.run
+      Savon.raise_errors = true
+    end
+
+    it "should return true if the request was successful" do
+      soap_response.should be_a_success
+    end
+
+    it "should return false if there was a SOAP fault" do
+      soap_fault_response.should_not be_a_success
+    end
+
+    it "should return false if there was an HTTP error" do
+      http_error_response.should_not be_a_success
     end
   end
 
@@ -31,12 +51,12 @@ describe Savon::SOAP::Response do
       Savon.raise_errors = true
     end
 
-    it "does not return true in case the response seems to be ok" do
-      new_response.soap_fault?.should be_false
+    it "should not return true in case the response seems to be ok" do
+      soap_response.soap_fault?.should be_false
     end
 
-    it "returns true in case of a SOAP fault" do
-      new_response(:body => ResponseFixture.soap_fault).soap_fault?.should be_true
+    it "should return true in case of a SOAP fault" do
+      soap_fault_response.soap_fault?.should be_true
     end
   end
 
@@ -47,14 +67,16 @@ describe Savon::SOAP::Response do
       Savon.raise_errors = true
     end
 
-    it "returns the SOAP fault message in case of a SOAP fault" do
-      new_response(:body => ResponseFixture.soap_fault).soap_fault.should ==
-        "(soap:Server) Fault occurred while processing."
+    it "should return a Savon::SOAP::Fault" do
+      soap_fault_response.soap_fault.should be_a(Savon::SOAP::Fault)
     end
 
-    it "returns the SOAP fault message in case of a SOAP 1.2 fault" do
-      new_response(:body => ResponseFixture.soap_fault12).soap_fault.should ==
-        "(soap:Sender) Sender Timeout"
+    it "should return a Savon::SOAP::Fault containing the HTTPI::Response" do
+      soap_fault_response.soap_fault.http.should be_an(HTTPI::Response)
+    end
+
+    it "should return a Savon::SOAP::Fault even if the SOAP response seems to be ok" do
+      soap_response.soap_fault.should be_a(Savon::SOAP::Fault)
     end
   end
 
@@ -65,12 +87,12 @@ describe Savon::SOAP::Response do
       Savon.raise_errors = true
     end
 
-    it "does not return true in case the response seems to be ok" do
-      new_response.http_error?.should_not be_true
+    it "should not return true in case the response seems to be ok" do
+      soap_response.http_error?.should_not be_true
     end
 
-    it "returns true in case of an HTTP error" do
-      new_response(:code => 500).http_error?.should be_true
+    it "should return true in case of an HTTP error" do
+      soap_response(:code => 500).http_error?.should be_true
     end
   end
 
@@ -81,46 +103,51 @@ describe Savon::SOAP::Response do
       Savon.raise_errors = true
     end
 
-    it "returns the HTTP error in case of an HTTP error" do
-      new_response(:code => 404, :body => "").http_error.should == "HTTP error (404)"
+    it "should return a Savon::HTTP::Error" do
+      http_error_response.http_error.should be_a(Savon::HTTP::Error)
     end
 
-    it "returns the HTTP error and response body (if available) in case of an HTTP error" do
-      new_response(:code => 404, :body => "Not found").http_error.should == "HTTP error (404): Not found"
+    it "should return a Savon::HTTP::Error containing the HTTPI::Response" do
+      http_error_response.http_error.http.should be_an(HTTPI::Response)
+    end
+
+    it "should return a Savon::HTTP::Error even if the HTTP response seems to be ok" do
+      soap_response.http_error.should be_a(Savon::HTTP::Error)
     end
   end
 
   describe "#to_hash" do
     it "should return the SOAP response body as a Hash" do
-      new_response.to_hash[:authenticate_response][:return].should ==
+      soap_response.to_hash[:authenticate_response][:return].should ==
         ResponseFixture.authentication(:to_hash)
-    end
-
-    it "should return a Hash for a SOAP multiRef response" do
-      response = new_response :body =>ResponseFixture.multi_ref
-      
-      response.to_hash[:list_response].should be_a(Hash)
-      response.to_hash[:multi_ref].should be_an(Array)
     end
   end
 
   describe "#to_xml" do
     it "should return the raw SOAP response body" do
-      new_response.to_xml.should == ResponseFixture.authentication
+      soap_response.to_xml.should == ResponseFixture.authentication
     end
   end
 
   describe "#http" do
-    it "should return the Net::HTTP response object" do
-      new_response.http.should be_an(HTTPI::Response)
+    it "should return the HTTPI::Response" do
+      soap_response.http.should be_an(HTTPI::Response)
     end
   end
 
-  def new_response(options = {})
+  def soap_response(options = {})
     defaults = { :code => 200, :headers => {}, :body => ResponseFixture.authentication }
     response = defaults.merge options
     
     Savon::SOAP::Response.new HTTPI::Response.new(response[:code], response[:headers], response[:body])
+  end
+
+  def soap_fault_response
+    soap_response :body => ResponseFixture.soap_fault
+  end
+
+  def http_error_response
+    soap_response :code => 404, :body => "Not found"
   end
 
 end